Typescript.Json

TSType

data TSType :: Nat -> IsObjType -> Type -> Type

data TSTypeF :: Nat -> IsObjType -> [Type] -> [Maybe Type] -> Type -> Type

data EnumLit

Constructors

ELString Text

ELNumber Scientific

data TSType_ p a

Constructors

forall k. TSType_

Fields

unTSType_ :: TSType p k a

data TSTypeF_ p as es b

Constructors

forall k. TSTypeF_

Fields

unTSTypeF_ :: TSTypeF p k as es b

Construction

Lists and Nullables

tsList :: TSType_ p a -> TSType p 'NotObj [a]

tsVector :: Vector v a => TSType_ p a -> TSType p 'NotObj (v a)

tsIsList :: IsList l => TSType_ p (Item l) -> TSType p 'NotObj l

Object

newtype ObjectProps p a b

A type aggregating key-value pairs for an object. Meant to be assembled using keyVal (for required properties) and keyValMay (for optional properties) and combined using its Applicative instance. To finally turn one into a TSType, use tsObject.

In a ObjectProps p a b, the a represents the overall aggregate type, and the b represents the type of the part that this ObjectProps is describing.

data MyType = MyType
  { mta :: Int
  , mtb :: Bool
  , mtc :: Maybe String
  }

myTypeProps :: ObjectProps p MyType MyType
myTypeProps = MyType
  <$> keyVal True mta "mta" (TSType_ tsBoundedInteger)
  <*> keyVal True mtb "mtb" (TSType_ tsBoolean)
  <*> keyValMay mtc "mtc" (TSType_ tsString)

myType :: TSType p 'IsObj MyType
myType = tsObject myTypeProps

ppType myType
-- => { mta: number, mtb: boolean, mtc?: string }

In the above, keyVal True mta "mta" tsBoundedInteger has the type ObjectProps p MyType Int, showing that it refers to the Int field of the MyType. The trick to using this is to assemble ObjectProps together using Applicative combinators until the a and b "match", and the ObjectProps describes the entire value. Then you can use tsObject.

Constructors

ObjectProps

Fields

getObjectProps :: Ap (Pre a (ObjMember (TSType_ p))) b

keyVal

Arguments

:: Bool
turn nullable types into optional params if possible

-> (a -> b)
project this pair's value out of the aggregate type

-> Mutability

-> Text
key (property name)

-> TSType_ p b

-> ObjectProps p a b

Create a single key-value pair for an object. If the first argument is True, will try to turn any nullable value into an optional property with non-nullable type if possible. Otherwise, always uses a required property.

keyValMay

Arguments

:: (a -> Maybe b)
project this pair's value out of the aggregate type, potentially revealing it is not present.

-> Mutability

-> Text
key (property name)

-> TSType_ p b

-> ObjectProps p a (Maybe b)

Create a single optional key-value pair for an object.

tsObject :: ObjectProps p a a -> TSType p 'IsObj a

Gather together object properties into a TSType. See ObjectProps for details on how to use this.

Tuple

newtype TupleVals p a b

A type aggregating values in a tuple type. Meant to be assembled using tupleVal and combined using its Applicative instance. To finally turn one into a TSType, use tsTuple.

In a TupleVals p a b, the a represents the overall aggregate type, and the b represents the type of the part that this TupleVals is describing.

data MyType = MyType
  { mta :: Int
  , mtb :: Bool
  , mtc :: String
  }

myTypeVals :: TupleVals p MyType MyType
myTypeVals = MyType
  <$> tupleVal mta (TSType_ tsBoundedInteger)
  <*> tupleVal mtb (TSType_ tsBoolean)
  <*> tupleVal mtc (TSType_ tsString)

myType :: TSType p 'NotObj MyType
myType = tsTuple myTypeVals

ppType myType
-- => [number, boolean, string]

In the above, tupleVal mta tsBoundedInteger has the type TupleVals p MyType Int, showing that it refers to the Int field of the MyType. The trick to using this is to assemble TupleVals together using Applicative combinators until the a and b "match", and the TupleVals describes the entire value. Then you can use tsTuple.

Note that the order that the Applicative combination matters: it determines the ordering of the tuple.

Constructors

TupleVals

Fields

getTupleVals :: Ap (Pre a (TSType_ p)) b

tupleVal

Arguments

:: (a -> b)
project this pair's value out of the aggregate type

-> TSType_ p b

-> TupleVals p a b

Create a singleton TupleVals, to be combined with applicative combinators with others.

tsTuple :: TupleVals p a a -> TSType p 'NotObj a

Gather together tuple values into a TSType. See TupleVals for details on how to use this.

stripObjectVals :: ObjectProps p a b -> TupleVals p a b

Unions

tsUnion :: TSUnionBranches p a -> TSType p 'NotObj a

Build up a union type from a unionBranch.

tsUnions :: (a -> NS I (b ': bs)) -> NP (Op a) (b ': bs) -> NP (TSType_ p) (b ': bs) -> TSType p 'NotObj a

Tagged

tagVal

Arguments

:: Mutability
readonly

-> Text
tag key

-> Text
tag value

-> ObjectProps p a ()

A utility for a simple situation of a "tag" key-value pair, where the property value is just a string literal singleton. Often used to simulate discriminated unions in typescript.

ppType . TSType_ . tsObject $
  tagVal "tag" "something"
-- => { tag: "sometning" }

You can combine this with other properties:

ppType . TSType_ . tsObject $
     tagVal "tag" "something"
  *> keyVal True "contents" id tsBoolean
-- => { tag: "someting", contents: boolean }

See also taggedObject, which uses an intersection instead of joining the keys directly.

taggedObject

Arguments

:: Mutability
readonly

-> Text
tag key

-> Text
tag value

-> TSType p 'IsObj a
contents (object)

-> TSType p 'IsObj a

A utility for a simple situation of a "tag" key-value pair intersected with an object type. Often used to simulate discriminated unions in typescript.

ppType . TSType_ $
     taggedObject "tag" "something" $
         tsObject $
           (,) $ keyVal True "name" fst (TSType_ tsText)
               * keyVal True "age" snd (TSType_ tsBoundedInteger)
-- => { tag: "something" } & { name: string, age: number }

See also taggedObject, which uses an intersection instead of joining the keys directly.

taggedValue

Arguments

:: TaggedValueOpts

-> Text
tag value

-> TSType_ p a
contents type

-> TSType p 'IsObj a

High-level utility to wrap a TSType_ with a "tag".

Assuming both flags are set to True:

If the contents type is an object, this will be { tagkey: tagvalue } & contents (the same behavior as taggedObject).

If the contents type is a nullable value, this well be { tagkey: tagvalue, contentskey?: contents}

Otherwise, this well be { tagkey: tagvalue, contentskey: contents}

If the first argument is False, will avoid case (1). If the second argument is False, will avoid case (2).

case1 :: TSType_ p (Text, Int)
case1 = TSType_ . tsObject $
  (,) <$> keyVal True "name" fst (TSType_ tsText)
      <*> keyVal True "age" snd (TSType_ tsBoundedInteger)

case2 :: TSType_ p (Maybe Int)
case2 = TSType_ $ tsNullable (TSType_ tsBoundedInteger)

case3 :: TSType_ p String
case3 = TSType_ tsString

ppType . TSType_ $ taggedValue True True "tag" "something" "contents" case1
-- => { tag: "something" } & { name: string, age: number }

ppType . TSType_ $ taggedValue True True "tag" "something" "contents" case2
-- => { tag: "something", contents?: number }

ppType . TSType_ $ taggedValue True True "tag" "something" "contents" case3
-- => { tag: "something", contents: string }

taggedNullary :: TaggedValueOpts -> Text -> a -> TSType_ p a

newtype TaggedBranches p a

A high-level data type describing the common pattern of a "tagged" union (sum types in Haskell), where each branch comes in the form of an object with a "tag" property with a string literal singleton, and the rest of the object is the contents. We would store an Either Int Bool as, say, { tag: Left, contents: number } | { tag: Right, contents: boolean }.

Meant to be constructed using taggedBranch and other Decide combinators.

Constructors

TaggedBranches

Fields

getTaggedBranches :: DecAlt1 (Branch p) a

data Branch p a

Constructors

Branch

Fields

branchTag :: Text

branchType :: Lift (TSType_ p) a

data TaggedValueOpts

Constructors

TVOTagAndContents TagAndContents

TVOTagIsKey TagIsKey

taggedBranch

Arguments

:: Text
Tag value

-> TSType_ p a

-> TaggedBranches p a

Create a singleton TaggedBranches, to be combined with Decide combinators with others. Can also be used with tsUnions if you want to combine a large number.

mergeTB :: (a -> Either b c) -> (b -> a) -> (c -> a) -> TaggedBranches p b -> TaggedBranches p c -> TaggedBranches p a

emptyTaggedBranch

Arguments

:: Text
Tag value

-> a
Pure "singleton"/literal rep value (ignored, just use ())

-> TaggedBranches p a

tsTaggedUnion :: TaggedValueOpts -> TaggedBranches p a -> TSType p 'NotObj a

tsTaggedUnions :: TaggedValueOpts -> (a -> NS I (b ': bs)) -> NP (Op a) (b ': bs) -> NP (TaggedBranches p) (b ': bs) -> TSType p 'NotObj a

Intersections

newtype IntersectVals p a b

A type aggregating the parts of an intersection. Meant to be assembled using intersectVal and combined using its Applicative instance. To finally turn one into a TSType, use tsIntersection.

In a IntersectVals p a b, the a represents the overall aggregate type, and the b represents the type of the part that this IntersectVals is describing.

data MyType = MyType
  { mta :: Int
  , mtb :: Bool
  , mtc :: Maybe String
  }

myType :: TSType p 'IsObj MyType
myType = tsObject $ MyType
  <$> keyVal True mta "mta" (TSType_ tsBoundedInteger)
  <*> keyVal True mtb "mtb" (TSType_ tsBoolean)
  <*> keyValMay mtc "mtc" (TSType_ tsString)

-- { tag: "something" }
tagType :: TSType p 'IsObj ()
tagType = tagVal "tag" "something"

myTaggedType :: IntersectVals p MyType MyType
myTaggedType = intersectVal tagType
            .> intersectVal myType

ppType $ tsIntersection myTaggedType
-- => { tag: "something" } &  { mta: number, mtb: boolean, mtc?: string }

This works in the same style as TupleVals and ObjectProps, so see those types for more examples of using an Applicative instance.

If any of the objects in the intersection have duplicate keys of the same type, the property value from the rightmost branch will be used when decoding.

If any of the objects in the intersection have duplicate keys of different types, this describes an invalid typescript type. The behavior of encoding/decoding is undefined; for encoding, the result will most likely not typecheck in typescript. for decoding, the result will most likely fail to parse.

Constructors

IntersectVals

Fields

getIntersectVals :: Ap1 (Pre a (TSType p 'IsObj)) b

intersectVal :: (a -> b) -> TSType p 'IsObj b -> IntersectVals p a b

Create a singleton IntersectVals, to be combined with applicative combinators with others.

Note that the input type must an object literal, indicated by 'IsObj

tsIntersection :: IntersectVals p a a -> TSType p 'IsObj a

Gather together intersection values into a TSType. See IntersectionVals for details on how to use this.

Named

tsNamed :: Text -> TSType p k a -> TSNamed p k '[] '[] a

Wrap a type in a name, in a way that preserves k (whether or not the type is an object literal).

tsNamed_ :: Text -> TSType_ p a -> TSNamed_ p '[] '[] a

Wrap a type in a name.

When printing this type, only the name will appear. However, when generating exports (typeExports, typeExportsF), these named types will be extracted and de-duplicated to the top level.

If the same name appears twice within a type, it must describe the same type underneath. Otherwise, behavior is undefined.

Generics

tsGeneric

Arguments

:: Text
Name

-> NP2 (Param p) as es
Name of parameters and extends

-> (forall r. SNat_ r -> NP2 (Arg_ (Plus r p)) as es -> TSType (Plus r p) k b)
Type function

-> TSNamed p k as es b

A parameterized type with multiple parameters.

tsGeneric1

Arguments

:: Text
Name of the type

-> Param p a e
Name of the parameter (used for printing) and what it extends

-> (forall r. SNat_ r -> Arg_ (Plus r p) a e -> TSType (Plus r p) k b)
Make a type, given the type parameter

-> TSNamed p k '[a] '[e] b

Create a single-argument generic (parameterized) type.

For example, we could make a type imitating Maybe in Haskell:

mkMaybe :: TSType_ p a -> TSType_ p (Maybe a)
mkMaybe = tsUnion $ decide (maybe (Left ()) Right) $
    (tsTagged

tsGeneric2 :: Text -> Param p a ea -> Param p b eb -> (forall r. SNat_ r -> Arg_ (Plus r p) a ea -> Arg_ (Plus r p) b eb -> TSType (Plus r p) k c) -> TSNamed p k '[a, b] '[ea, eb] c

tsGeneric3 :: Text -> Param p a ea -> Param p b eb -> Param p c ec -> (forall r. SNat_ r -> Arg_ (Plus r p) a ea -> Arg_ (Plus r p) b eb -> Arg_ (Plus r p) c ec -> TSType (Plus r p) k d) -> TSNamed p k '[a, b, c] '[ea, eb, ec] d

tsGenericInterface

Arguments

:: Text
name

-> NP2 (Param p) as es
Name of parameters and extends

-> Lift (TSAppliedF p 'IsObj as es) a
if it extends any other object

-> (forall r. SNat_ r -> NP2 (Arg_ (Plus r p)) as es -> TSKeyVal (Plus r p) b)

-> TSNamed p 'IsObj as es (a, b)

tsApplied1 :: TSNamed p k '[a] '[e] b -> Arg_ p a e -> TSType p k b

tsApplied2 :: TSNamed p k '[a, b] '[ea, eb] c -> Arg_ p a ea -> Arg_ p b eb -> TSType p k c

tsApplied3 :: TSNamed p k '[a, b, c] '[ea, eb, ec] d -> Arg_ p a ea -> Arg_ p b eb -> Arg_ p c ec -> TSType p k d

tsApplied :: TSNamed p k as es b -> NP2 (Arg_ p) as es -> TSType p k b

tsApply1

Arguments

:: TSTypeF p k '[a] '[e] b
type function

-> Arg_ p a e
thing to apply

-> TSType p k b

tsApply2

Arguments

:: TSTypeF p k '[a, b] '[ea, eb] c
type function

-> Arg_ p a ea
thing to apply

-> Arg_ p b eb
thing to apply

-> TSType p k c

tsApply3

Arguments

:: TSTypeF p k '[a, b, c] '[ea, eb, ec] d
type function

-> Arg_ p a ea
thing to apply

-> Arg_ p b eb
thing to apply

-> Arg_ p c ec
thing to apply

-> TSType p k d

tsApply

Arguments

:: TSTypeF p k as es b
type function

-> NP2 (Arg_ p) as es
thing to apply

-> TSType p k b

Primitives

tsBoolean :: TSType p 'NotObj Bool

tsNumber :: TSType p 'NotObj Scientific

tsBoundedInteger :: (Integral a, Bounded a) => TSType p 'NotObj a

tsInteger :: Integral a => TSType p 'NotObj a

tsRealFloat :: RealFloat a => TSType p 'NotObj a

tsDouble :: TSType p 'NotObj Double

tsBigInt :: TSType p 'NotObj Integer

tsText :: TSType p 'NotObj Text

tsLazyText :: TSType p 'NotObj Text

tsString :: TSType p 'NotObj String

tsStringLit :: Text -> TSType p 'NotObj ()

tsNumericLit :: Scientific -> TSType p 'NotObj ()

tsIntegerLit :: Integral a => a -> TSType p 'NotObj ()

tsBigIntLit :: Integer -> TSType p 'NotObj ()

tsUnknown :: TSType p 'NotObj Value

tsAny :: TSType p 'NotObj Value

tsVoid :: TSType p 'NotObj ()

tsUndefined :: TSType p 'NotObj ()

tsNull :: TSType p 'NotObj ()

tsNever :: TSType p 'NotObj Void

tsMaybe

Arguments

:: Text
the "nothing" constructor

-> Text
the "just" field

-> TSType_ p a

-> TSType p 'NotObj (Maybe a)

Builidng Enum

With Vector

data FinIso n a

Constructors

FinIso

Fields

fiGet :: Fin n -> a

fiPut :: a -> Fin n

tsEnumWith :: Text -> FinIso n a -> Vec n (Text, EnumLit) -> TSNamed p 'NotObj '[] '[] a

tsIntEnumFrom :: Text -> Int -> FinIso n a -> Vec n Text -> TSNamed p 'NotObj '[] '[] a

tsIntEnum :: Text -> FinIso n a -> Vec n Text -> TSNamed p 'NotObj '[] '[] a

With Map

data EnumMap a

Constructors

EnumMap

Fields

emGet :: Map EnumLit (Text, a)

emPut :: a -> EnumLit

tsEnumMap :: Text -> EnumMap a -> TSNamed p 'NotObj '[] '[] a

With Enum

tsFinEnum :: Enum a => Text -> Vec (EnumSize a) (Text, EnumLit) -> TSNamed p 'NotObj '[] '[] a

tsFinIntEnum :: Enum a => Text -> Vec (EnumSize a) Text -> TSNamed p 'NotObj '[] '[] a

Printing

ppType :: TSType 'Z k a -> Doc x

ppType_ :: TSType_ 'Z a -> Doc x

ppNamed :: TSNamed 'Z k as es a -> Doc x

ppNamed_ :: TSNamed_ 'Z as es a -> Doc x

typeExports' :: Vec p Text -> TSType p k a -> Doc x

typeExports :: TSType 'Z k a -> Doc x

typeExports_ :: TSType_ 'Z a -> Doc x

namedTypeExports' :: Vec p Text -> TSNamed p k as es a -> Doc x

namedTypeExports :: TSNamed 'Z k as es a -> Doc x

namedTypeExports_ :: TSNamed_ 'Z as es a -> Doc x

Serializing

encodeType :: TSType 'Z k a -> a -> ByteString

encodeTypeStrict :: TSType 'Z k a -> a -> ByteString

typeToValue :: TSType 'Z k a -> a -> Value

Parsing

decodeType :: TSType 'Z k a -> ByteString -> Either (ParseError ParseErr) a

decodeTypeStrict :: TSType 'Z k a -> ByteString -> Either (ParseError ParseErr) a

parseType :: forall k a. TSType 'Z k a -> Parse a