Typescript.Json.Core.Assign

Documentation

isAssignable :: TSType 'Z k a -> TSType 'Z j b -> Bool

answers: is X assignable to Y?

https://basarat.gitbook.io/typescript/type-system/type-compatibility

reAssign :: TSType 'Z k a -> TSType 'Z j b -> Maybe (Assign a b)

If X is assignable to Y, then convert x to the more general y, potentially losing information.

unsafeReAssign :: TSType 'Z k a -> TSType 'Z j b -> Assign a b

unsafeAssign :: Assign a b

Completely 100% unsafe, but at least it doesn't require having a type without any free variables