Typescript.Json.Core.Print

Documentation

ppType :: TSType 'Z k a -> Doc x

ppType_ :: TSType_ 'Z a -> Doc x

ppEnumLit :: EnumLit -> Doc x

ppNamedBase :: Text -> TSNamedBase a -> Doc x

ppNamed :: TSNamed 'Z k as es a -> Doc x

ppNamed_ :: TSNamed_ 'Z as es a -> Doc x

ppNamed' :: Vec p Text -> TSNamed p k as es a -> Doc x

typeExports_ :: TSType_ 'Z a -> Doc x

typeExports :: TSType 'Z k a -> Doc x

typeExports' :: Vec p Text -> TSType p k a -> Doc x

namedTypeExports_ :: TSNamed_ 'Z as es a -> Doc x

namedTypeExports :: TSNamed 'Z k as es a -> Doc x

namedTypeExports' :: Vec p Text -> TSNamed p k as es a -> Doc x

ppMap :: forall x. Map Text (Set Text, Doc x) -> Doc x

flattenNamedType :: Vec p Text -> TSNamed p k as es a -> Map Text (Set Text, Doc x)

Pull out all of the named types to be top-level type declarations, and have create a map of all of those declarations.

flattenType :: Vec p Text -> TSType p k a -> Map Text (Set Text, Doc x)

Ignores the top level type, so why even bother?

flattenType_

Arguments

:: forall p k a x. Vec p Text

-> Set Text
types it is currently named under

-> TSType p k a

-> State (Map Text (Set Text, Doc x)) (Set Text)

TODO: stop if it is a recursive reference