Typescript.Json.Generics

Typeclass-based

class GTSType (f :: Type -> Type) where

Methods

gtoTSType :: TSOpts -> NP (TSType_ p) (LeafTypes f) -> TSType_ p (f x)

data TSOpts

Constructors

TSOpts

Fields

tsoFieldModifier :: String -> Text

tsoPreserveMaybe :: Bool

If False, turns all immediate Maybe fields into optional fields. If True, keeps the field required.

tsoMakeMaybe :: MakeMaybe

Turn a a into a Maybe a, used to fill in top-level Maybe fields or if tsoPreserveMaybe is True.

tsoCollapseNullable :: Bool

When assigning fields, if a type is nullable (null is assignable to it, or it's a union where null or undefined is a member) then turn it into an optional field and strips the type of nulls. Note that this will inline any named definitions if the type is nullable.

tsoConstructorModifier :: String -> Text

tsoSumOpts :: TaggedValueOpts

tsoReadOnlyFields :: Mutability

tsoTypeNameModifier :: SomeTypeRep -> Text

class GTSSum (f :: Type -> Type) where

Methods

gtsSum :: TSOpts -> NP (TSType_ p) (LeafTypes f) -> TaggedBranches p (f x)

class (forall p b. Applicative (t p b)) => GTSProduct (t :: Nat -> Type -> Type -> Type) (f :: Type -> Type) | f -> t where

Methods

gtsProduct :: TSOpts -> NP (TSType_ p) (LeafTypes f) -> (a -> f x) -> t p a (f x)

class GTSTypeF (f :: Type -> Type) where

Methods

gtoTSTypeF :: TSOpts -> NP (TSType_ p) (LeafTypes f) -> TSTypeF_ p '[a] '['Nothing] (f a)

class GTSSumF (f :: Type -> Type) where

Methods

gtsSumF :: TSOpts -> NP (TSType_ p) (LeafTypes f) -> TSType p k a -> TaggedBranches p (f a)

class (forall p b. Applicative (t p b)) => GTSProductF (t :: Nat -> Type -> Type -> Type) (f :: Type -> Type) | f -> t where

Methods

gtsProductF :: TSOpts -> NP (TSType_ p) (LeafTypes f) -> (b -> f a) -> TSType p k a -> t p b (f a)

class GTSEnum (f :: Type -> Type) where

Methods

gtsNamedEnum :: EnumOpts -> TSNamed p 'NotObj '[] '[] (f x)

data EnumOpts

Constructors

EnumOpts

Fields

eoSelector :: String -> Text

eoLiteral :: String -> Maybe EnumLit

if Nothing, picked from free integers, starting from 0

newtype MakeMaybe

Constructors

MakeMaybe

Fields

runMakeMaybe :: forall p a. TSType_ p a -> TSType_ p (Maybe a)

simpleMakeMaybe

Arguments

:: Text
"none" key

-> Text
"some" key

-> MakeMaybe

"none" | { "some": x }

Default instances

class ToTSType a where

Minimal complete definition

Nothing

Methods

toTSType :: TSType_ p a

default toTSType :: (Generic a, GTSType (MRep a), All ToTSType (LeafTypes (MRep a))) => TSType_ p a

genericToTSType :: forall a p. (Generic a, GTSType (MRep a)) => TSOpts -> NP (TSType_ p) (LeafTypes (MRep a)) -> TSType_ p a

genericToTSType_ :: forall a p. (Generic a, GTSType (MRep a), All ToTSType (LeafTypes (MRep a))) => TSOpts -> TSType_ p a

genericToTSNamed :: forall a p. (Typeable a, Generic a, GTSType (MRep a)) => TSOpts -> NP (TSType_ p) (LeafTypes (MRep a)) -> TSNamed_ p '[] '[] a

genericToTSNamed_ :: forall a p. (Typeable a, Generic a, GTSType (MRep a), All ToTSType (LeafTypes (MRep a))) => TSOpts -> TSNamed_ p '[] '[] a

genericToTSType1 :: forall f p a. (Typeable f, Generic1 f, GTSTypeF (MRep1 f)) => TSOpts -> NP (TSType_ p) (LeafTypes (MRep1 f)) -> TSType_ p a -> TSType_ p (f a)

genericToTSType1_ :: forall f p a. (Typeable f, Generic1 f, GTSTypeF (MRep1 f), All ToTSType (LeafTypes (MRep1 f))) => TSOpts -> TSType_ p a -> TSType_ p (f a)

genericToTSTypeF :: forall f p a. (Generic1 f, GTSTypeF (MRep1 f)) => TSOpts -> NP (TSType_ p) (LeafTypes (MRep1 f)) -> TSTypeF_ p '[a] '['Nothing] (f a)

genericToTSTypeF_ :: forall f p a. (Generic1 f, GTSTypeF (MRep1 f), All ToTSType (LeafTypes (MRep1 f))) => TSOpts -> TSTypeF_ p '[a] '['Nothing] (f a)

genericToTSNamedF :: forall f p a. (Typeable f, Generic1 f, GTSTypeF (MRep1 f)) => TSOpts -> NP (TSType_ p) (LeafTypes (MRep1 f)) -> TSNamed_ p '[a] '['Nothing] (f a)

genericToTSNamedF_ :: forall f p a. (Typeable f, Generic1 f, GTSTypeF (MRep1 f), All ToTSType (LeafTypes (MRep1 f))) => TSOpts -> TSNamed_ p '[a] '['Nothing] (f a)

genericNamedEnum :: forall a p. (Generic a, GTSEnum (Rep a)) => EnumOpts -> TSNamed p 'NotObj '[] '[] a

Util

type family (as :: [k]) ++ (bs :: [k]) :: [k] where ...

Equations

'[] ++ bs = bs

(a ': as) ++ bs = a ': (as ++ bs)

splitNP :: NP p as -> NP f (as ++ bs) -> (NP f as, NP f bs)

mfrom :: forall a x. Generic a => a -> MRep a x

mto :: forall a x. Generic a => MRep a x -> a

type MRep a = PullMaybe (Rep a)

type MRep1 a = PullMaybe (Rep1 a)

type family PullMaybe (f :: Type -> Type) :: Type -> Type where ...

Equations

PullMaybe (K1 i (Maybe a)) = KM1 i a

PullMaybe (K1 i a) = K1 i a

PullMaybe U1 = U1

PullMaybe V1 = V1

PullMaybe (M1 x y f) = M1 x y (PullMaybe f)

PullMaybe Par1 = Par1

PullMaybe (f :+: g) = PullMaybe f :+: PullMaybe g

PullMaybe (f :*: g) = PullMaybe f :*: PullMaybe g

newtype KM1 i a x

Constructors

KM1

Fields

unKM1 :: Maybe a

data FooBar a

Constructors

FooBar Foo Bar a

data List a

Constructors

LNil

LCons a (List a)