Typescript.Json.Types

Documentation

data TSPrim :: Type -> Type where

Constructors

TSNumber :: TSPrim Scientific

TSBigInt :: TSPrim Integer

TSString :: TSPrim Text

TSUnknown :: TSPrim Value

TSAny :: TSPrim Value

data TSNamedBase :: Type -> Type where

Named primitive types, that cannot be anonymous

Constructors

TSEnum :: Vec n (Text, EnumLit) -> TSNamedBase (Fin n)

data TSBase :: Type -> Type where

Constructors

TSBoolean :: TSBase Bool

TSStringLit :: Text -> TSBase ()

TSNumericLit :: Scientific -> TSBase ()

TSBigIntLit :: Integer -> TSBase ()

TSVoid :: TSBase ()

TSUndefined :: TSBase ()

TSNull :: TSBase ()

TSNever :: TSBase Void

data EnumLit

Constructors

ELString Text

ELNumber Scientific

data TSType :: Nat -> IsObjType -> Type -> Type where

Constructors

TSArray :: ILan [] (TSType p k) a -> TSType p 'NotObj a

TSTuple :: PreT Ap (TSType_ p) a -> TSType p 'NotObj a

TSObject :: TSKeyVal p a -> TSType p 'IsObj a

TSSingle :: TSType p 'IsObj a -> TSType p 'NotObj a

TSUnion :: TSUnionBranches p a -> TSType p 'NotObj a

TSNamedType :: TSApplied p k a -> TSType p k a

TSVar :: !(Fin p) -> TSType p 'NotObj a

TSIntersection :: PreT Ap1 (TSType p 'IsObj) a -> TSType p 'IsObj a

TSTransformType :: ICoyoneda (TSTransform p k) a -> TSType p k a

TSPrimType :: PS TSPrim a -> TSType p 'NotObj a

TSBaseType :: ICoyoneda TSBase a -> TSType p 'NotObj a

data TSType_ p a

Constructors

forall k. TSType_

Fields

unTSType_ :: TSType p k a

data TSNamed p k as es a

Constructors

TSNamed

Fields

tsnName :: Text

tsnType :: TSNameable p k as es a

data TSNamed_ p as es a

Constructors

forall k. TSNamed_ (TSNamed p k as es a)

withTSNamed_ :: (forall k. TSNamed p k as es a -> r) -> TSNamed_ p as es a -> r

newtype Assign a b

Constructors

Assign

Fields

runAssign :: a -> Either Text b

data TSNameable :: Nat -> IsObjType -> [Type] -> [Maybe Type] -> Type -> Type where

Constructors

TSNFunc :: TSTypeF p k as es a -> TSNameable p k as es a

TSNBaseType :: ICoyoneda TSNamedBase a -> TSNameable p 'NotObj '[] '[] a

data Param p a b

Constructors

Param

Fields

paramName :: Text

paramExtends :: MP (TSType_ p) b

data Arg p k a b

Constructors

Arg

Fields

argType :: TSType p k a

argAssign :: MP (Assign a) b

data Arg_ p a b

Constructors

forall k. Arg_ (Arg p k a b)

withArg_ :: (forall k. Arg p k a e -> r) -> Arg_ p a e -> r

data ObjMember f a

Constructors

ObjMember

Fields

objMemberReadOnly :: Mutability

objMemberKey :: Text

objMemberVal :: (f :+: ILan Maybe f) a

data Mutability

Constructors

Mutable

ReadOnly

data TSTransform :: Nat -> IsObjType -> Type -> Type where

Built-in named type transformers in typescript, that should be displayed as such when printed.

Constructors

TSPartial :: TSType p 'IsObj a -> TSTransform p 'IsObj (Maybe a)

TSReadOnly :: TSType p 'IsObj a -> TSTransform p 'IsObj a

TSPickPartial :: TSType p 'IsObj a -> TSType p 'NotObj (KeySubset ks a) -> Assign (KeySubset ks a) (KeyOf ks a) -> TSTransform p 'IsObj (Maybe a)

TSOmitPartial :: TSType p 'IsObj a -> TSType p 'NotObj (KeySubset ks a) -> Assign (KeySubset ks a) (KeyOf ks a) -> TSTransform p 'IsObj (Maybe a)

TSStringManipType :: TSStringManip -> TSType p 'NotObj a -> Assign a ExtendsString -> TSTransform p 'NotObj a

data TSStringManip

Constructors

TSUppercase

TSLowercase

TSCapitalize

TSUncapitalize

newtype ExtendsString

Constructors

ExtendsString

Fields

getExtendsString :: Text

extendsString :: TSType p 'NotObj ExtendsString

Canonical type to test against when using TSStringManipType

applyTransform :: TSTransform p k a -> TSType p k a

Transforms are just deferred evaluations of functions. here they are actually applied

type TSKeyVal p = PreT Ap (ObjMember (TSType_ p))

type TSUnionBranches p = DecAlt1 (TSType_ p)

mapTSType_ :: (forall k. TSType p k a -> TSType us k b) -> TSType_ p a -> TSType_ us b

withTSType_ :: (forall k. TSType p k a -> r) -> TSType_ p a -> r

onTSType_ :: (TSType p 'NotObj a -> r) -> (TSType p 'IsObj a -> r) -> TSType_ p a -> r

decideTSType_ :: TSType_ p ~> (TSType p 'NotObj :+: TSType p 'IsObj)

data IsObjType

Constructors

NotObj

IsObj

data SIsObjType :: IsObjType -> Type where

Constructors

SNotObj :: SIsObjType 'NotObj

SIsObj :: SIsObjType 'IsObj

data SNat_ :: Nat -> Type where

Constructors

SZ_ :: SNat_ 'Z

SS_ :: SNat_ n -> SNat_ ('S n)

interpretObjMember :: Invariant g => (Mutability -> Text -> f ~> g) -> (forall x. Mutability -> Text -> f x -> g (Maybe x)) -> ObjMember f ~> g

tsObjType :: TSType p k a -> SIsObjType k

collapseIsObj :: TSType p 'IsObj a -> TSKeyVal p a

splitKeyVal :: TSKeyVal p a -> Map Text (Some (Pre a (TSType_ p)))

isObjKeys :: TSType p 'IsObj a -> Set Text

mkNullable :: TSType p k a -> TSType p 'NotObj (Maybe a)

toVarArgs :: forall p as es. NP2 (Param p) as es -> (SNat_ (Length as), NP2 (Arg_ (Plus (Length as) p)) as es)

Generics

data TSTypeF :: Nat -> IsObjType -> [Type] -> [Maybe Type] -> Type -> Type where

Constructors

TSGeneric :: NP2 (Param p) as es -> (forall r. SNat_ r -> NP2 (Arg_ (Plus r p)) as es -> TSType (Plus r p) k b) -> TSTypeF p k as es b

TSGenericInterface

Fields

:: NP2 (Param p) as es

-> (a -> b -> c)

covariant day convolution: combine base and new

-> (c -> (a, b))

contravariant day convolution: split into base and new

-> Lift (TSAppliedF p 'IsObj as es) a

-> (forall r. SNat_ r -> NP2 (Arg_ (Plus r p)) as es -> TSKeyVal (Plus r p) b)

-> TSTypeF p 'IsObj as es c

data TSTypeF_ p as es b

Constructors

forall k. TSTypeF_

Fields

unTSTypeF_ :: TSTypeF p k as es b

data TSApplied p k a

Constructors

forall as es. (:$)

Fields

tsaFunc :: TSNamed p k as es a

tsaArgs :: NP2 (Arg_ p) as es

data TSAppliedF p k as es a

Constructors

forall bs ds. (:?)

Fields

tsafFunc :: TSNamed p k bs ds a

tsafArgs :: NP2 (ArgF_ p as es) bs ds

data ArgF p k as es a e

Constructors

ArgF

Fields

argfType :: TSTypeF p k as es a

argfAssign :: MP (Assign a) e

data ArgF_ p as es a e

Constructors

forall k. ArgF_ (ArgF p k as es a e)

withArgF_ :: (forall k. ArgF p k as es a e -> r) -> ArgF_ p as es a e -> r

mapTSTypeF_ :: (forall k. TSTypeF p k as es b -> TSTypeF q k as' es' b') -> TSTypeF_ p as es b -> TSTypeF_ q as' es' b'

withTSTypeF_ :: (forall k. TSTypeF p k as es b -> r) -> TSTypeF_ p as es b -> r

pattern Param' :: Text -> Param p a 'Nothing

pattern Arg' :: TSType p k a -> Arg p k a 'Nothing

data ParseErr

Constructors

PEInvalidEnum [(Text, EnumLit)]

PEInvalidString Text Text

PEInvalidNumber Scientific Scientific

PEInvalidBigInt Integer Integer

PEPrimitive (Some TSPrim) Text

PEExtraTuple Int Int

PENotInUnion (NonEmpty (Doc ()))

PENever

tsApply

Arguments

:: TSTypeF p k as es b
type function

-> NP2 (Arg_ p) as es
thing to apply

-> TSType p k b

tsApply1

Arguments

:: TSTypeF p k '[a] '[e] b
type function

-> Arg_ p a e
thing to apply

-> TSType p k b

tsApplyF :: forall p k as es bs ds b. TSTypeF p k as es b -> NP2 (ArgF_ p bs ds) as es -> NP2 (Arg_ p) bs ds -> TSType p k b

tsApplyVar :: forall p k as es b. TSTypeF p k as es b -> TSType (Plus (Length as) p) k b

Apply a TypeF with free variables

tsfParams :: TSTypeF p k as es b -> Vec (Length as) Text

tsShift :: forall r p k a. SNat_ r -> TSType p k a -> TSType (Plus r p) k a

shiftApplied :: SNat_ r -> TSApplied p k a -> TSApplied (Plus r p) k a

shiftAppliedF :: SNat_ r -> TSAppliedF p k as es a -> TSAppliedF (Plus r p) k as es a

shiftNameable :: SNat_ r -> TSNameable p k as es a -> TSNameable (Plus r p) k as es a

shiftTypeF :: forall r p k as es a. SNat_ r -> TSTypeF p k as es a -> TSTypeF (Plus r p) k as es a

isNullable :: TSType p k a -> Maybe (ILan Maybe (TSType p 'NotObj) a)

Pulls out the (x | y | z) from (x | y | z) | null. The result is an ILan where the item inside is non-nullable.

This removes ALL nulls and undefineds.

Note that, if the result is Just (it is nullable), this inlines all named times because it changes the structure of the type itself.

flattenUnion :: TSType p k a -> DecAlt1 (TSType_ p) a

singletonValue :: TSType p k a -> Maybe a

If the type is a singleton type with only one inhabitant, this gives that type.

summonValues :: TSType p k a -> [a]

Summon some example values

Key Subset

data KeyOf ks a

Constructors

KeyOf

Fields

getKeyOf :: NS SSym ks

newtype KeySubset ks a

Constructors

KeySubset

Fields

getKeySubset :: NP (Maybe :.: SSym) ks

keyOf :: TSType p 'IsObj a -> (forall ks. NP SSym ks -> TSType p 'NotObj (KeyOf ks a) -> r) -> r

data KeySubsetErr

Constructors

KSEUnmatched (NonEmpty Text)

KSENoMatches

mkKeySubset :: forall ks a p. Set Text -> NP SSym ks -> Either KeySubsetErr (TSType p 'NotObj (KeySubset ks a), Assign (KeySubset ks a) (KeyOf ks a))