Typescript.Json.Types.Combinators

Description

Invariant "Functor combinators" (a la Data.Functor.Combinator) used for type definitions.

Documentation

data PS f a

Give f an Invariant instance, except the post-map can be partial. PS stands for parser-and-serializer: With a PS f a, you can serialize an a using it (total), or you can parse an a out of it (partial, with a Text error).

Constructors

forall r. PS

Fields

psItem :: f r

psParser :: r -> Either Text a

psSerializer :: a -> r

data ILan g h a

If h a is some sort of witness on a, then ILan g h a essentially turns it into a witness on g a.

It's used twice in the library: with [], it turns F a, a serializerparser on a, into ILan [] F [a], a serializerparser on [a], a list of them. It also does the same thing with Maybe, for optional parsers.

For all of the functions here, it's generally more understandable when you substitute a specific functor in for g.

Construct using ilan, usually.

Constructors

forall x. ILan (g x -> a) (a -> g x) (h x)

ilan :: h a -> ILan g h (g a)

Given a witness h a, turn it into a witness h (g a) using ILan.

interpretILan :: Invariant f => (forall x. h x -> f (g x)) -> ILan g h a -> f a

Interpret ILan into any invariant context, within its "lifted" witness.

For example, for ILan [], its type is:

(forall x. h x -> f [x])
  -> ILan g h a
  -> f a

If you can interpret h x into a list of x in some context, then it interprets out the a.

interpretCoILan :: Functor f => (forall x. h x -> f (g x)) -> ILan g h a -> f a

Interpret ILan, but requiring only a Functor context.

interpretContraILan :: Contravariant f => (forall x. h x -> f (g x)) -> ILan g h a -> f a

Interpret ILan, but requiring only a Contravariant context.

data ICoyoneda f a

Invariant Coyoneda: the "free invariant functor". Gives any f and Invariant instance.

Constructors

forall r. ICoyoneda (a -> r) (r -> a) (f r)

icoToContraco :: ICoyoneda f ~> Coyoneda f

Strip out the covariant aspect of an invariant coyoneda to leave the contravariant part.

icoToCoco :: ICoyoneda f ~> Coyoneda f

Strip out the contravariant aspect of an invariant coyoneda to leave the covariant part.

data MP :: (k -> Type) -> Maybe k -> Type where

Constructors

MPNothing :: MP f 'Nothing

MPJust :: f a -> MP f ('Just a)

data NP2 :: (k -> j -> Type) -> [k] -> [j] -> Type where

Constructors

Nil2 :: NP2 f '[] '[]

(:**) :: f a b -> NP2 f as bs -> NP2 f (a ': as) (b ': bs) infixr 5

hmap2 :: forall f g as bs. (forall a b. f a b -> g a b) -> NP2 f as bs -> NP2 g as bs

htraverse2 :: forall f g h as bs. Applicative h => (forall a b. f a b -> h (g a b)) -> NP2 f as bs -> h (NP2 g as bs)

hfoldMap2 :: forall f m as bs. Monoid m => (forall a b. f a b -> m) -> NP2 f as bs -> m

np2Left :: forall f g as bs. (forall a b. f a b -> g a) -> NP2 f as bs -> NP g as

splitAp :: forall f b. Ap f b -> [Some f]

findNP :: forall f g as. (forall a. f a -> Maybe (g a)) -> NP f as -> Maybe (NS g as)