Typescript.Json.Union

Unions

tsUnion :: TSUnionBranches p a -> TSType p 'NotObj a

Build up a union type from a unionBranch.

tsUnions :: (a -> NS I (b ': bs)) -> NP (Op a) (b ': bs) -> NP (TSType_ p) (b ': bs) -> TSType p 'NotObj a

Tagged

tagVal

Arguments

:: Mutability
readonly

-> Text
tag key

-> Text
tag value

-> ObjectProps p a ()

A utility for a simple situation of a "tag" key-value pair, where the property value is just a string literal singleton. Often used to simulate discriminated unions in typescript.

ppType . TSType_ . tsObject $
  tagVal "tag" "something"
-- => { tag: "sometning" }

You can combine this with other properties:

ppType . TSType_ . tsObject $
     tagVal "tag" "something"
  *> keyVal True "contents" id tsBoolean
-- => { tag: "someting", contents: boolean }

See also taggedObject, which uses an intersection instead of joining the keys directly.

taggedObject

Arguments

:: Mutability
readonly

-> Text
tag key

-> Text
tag value

-> TSType p 'IsObj a
contents (object)

-> TSType p 'IsObj a

A utility for a simple situation of a "tag" key-value pair intersected with an object type. Often used to simulate discriminated unions in typescript.

ppType . TSType_ $
     taggedObject "tag" "something" $
         tsObject $
           (,) $ keyVal True "name" fst (TSType_ tsText)
               * keyVal True "age" snd (TSType_ tsBoundedInteger)
-- => { tag: "something" } & { name: string, age: number }

See also taggedObject, which uses an intersection instead of joining the keys directly.

taggedValue

Arguments

:: TaggedValueOpts

-> Text
tag value

-> TSType_ p a
contents type

-> TSType p 'IsObj a

High-level utility to wrap a TSType_ with a "tag".

Assuming both flags are set to True:

If the contents type is an object, this will be { tagkey: tagvalue } & contents (the same behavior as taggedObject).

If the contents type is a nullable value, this well be { tagkey: tagvalue, contentskey?: contents}

Otherwise, this well be { tagkey: tagvalue, contentskey: contents}

If the first argument is False, will avoid case (1). If the second argument is False, will avoid case (2).

case1 :: TSType_ p (Text, Int)
case1 = TSType_ . tsObject $
  (,) <$> keyVal True "name" fst (TSType_ tsText)
      <*> keyVal True "age" snd (TSType_ tsBoundedInteger)

case2 :: TSType_ p (Maybe Int)
case2 = TSType_ $ tsNullable (TSType_ tsBoundedInteger)

case3 :: TSType_ p String
case3 = TSType_ tsString

ppType . TSType_ $ taggedValue True True "tag" "something" "contents" case1
-- => { tag: "something" } & { name: string, age: number }

ppType . TSType_ $ taggedValue True True "tag" "something" "contents" case2
-- => { tag: "something", contents?: number }

ppType . TSType_ $ taggedValue True True "tag" "something" "contents" case3
-- => { tag: "something", contents: string }

taggedNullary :: TaggedValueOpts -> Text -> a -> TSType_ p a

newtype TaggedBranches p a

A high-level data type describing the common pattern of a "tagged" union (sum types in Haskell), where each branch comes in the form of an object with a "tag" property with a string literal singleton, and the rest of the object is the contents. We would store an Either Int Bool as, say, { tag: Left, contents: number } | { tag: Right, contents: boolean }.

Meant to be constructed using taggedBranch and other Decide combinators.

Constructors

TaggedBranches

Fields

getTaggedBranches :: DecAlt1 (Branch p) a

data Branch p a

Constructors

Branch

Fields

branchTag :: Text

branchType :: Lift (TSType_ p) a

data TaggedValueOpts

Constructors

TVOTagAndContents TagAndContents

TVOTagIsKey TagIsKey

taggedBranch

Arguments

:: Text
Tag value

-> TSType_ p a

-> TaggedBranches p a

Create a singleton TaggedBranches, to be combined with Decide combinators with others. Can also be used with tsUnions if you want to combine a large number.

mergeTB :: (a -> Either b c) -> (b -> a) -> (c -> a) -> TaggedBranches p b -> TaggedBranches p c -> TaggedBranches p a

emptyTaggedBranch

Arguments

:: Text
Tag value

-> a
Pure "singleton"/literal rep value (ignored, just use ())

-> TaggedBranches p a

tsTaggedUnion :: TaggedValueOpts -> TaggedBranches p a -> TSType p 'NotObj a

tsTaggedUnions :: TaggedValueOpts -> (a -> NS I (b ': bs)) -> NP (Op a) (b ': bs) -> NP (TaggedBranches p) (b ': bs) -> TSType p 'NotObj a

Misc

tsMaybe

Arguments

:: Text
the "nothing" constructor

-> Text
the "just" field

-> TSType_ p a

-> TSType p 'NotObj (Maybe a)