Typescript.Json

TSType

data TSType :: Nat -> IsObjType -> Type -> Type

data TSTypeF :: Nat -> IsObjType -> [Type] -> [Maybe Type] -> Type -> Type

data EnumLit

Constructors

ELString Text

ELNumber Scientific

data TSType_ p a

Constructors

forall k. TSType_

Fields

unTSType_ :: TSType p k a

data TSTypeF_ p as es b

Constructors

forall k. TSTypeF_

Fields

unTSTypeF_ :: TSTypeF p k as es b

Construction

module Typescript.Json.Primitive

module Typescript.Json.Collection

module Typescript.Json.Union

module Typescript.Json.Enum

module Typescript.Json.Parameterized

Named

tsNamed :: Text -> TSType p k a -> TSNamed p k '[] '[] a

Wrap a type in a name, in a way that preserves k (whether or not the type is an object literal).

tsNamed_ :: Text -> TSType_ p a -> TSNamed_ p '[] '[] a

Wrap a type in a name.

When printing this type, only the name will appear. However, when generating exports (typeExports, typeExportsF), these named types will be extracted and de-duplicated to the top level.

If the same name appears twice within a type, it must describe the same type underneath. Otherwise, behavior is undefined.

Printing

ppType :: TSType 'Z k a -> Doc x

ppType_ :: TSType_ 'Z a -> Doc x

ppNamed :: TSNamed 'Z k as es a -> Doc x

ppNamed_ :: TSNamed_ 'Z as es a -> Doc x

typeExports' :: Vec p Text -> TSType p k a -> Doc x

typeExports :: TSType 'Z k a -> Doc x

typeExports_ :: TSType_ 'Z a -> Doc x

namedTypeExports' :: Vec p Text -> TSNamed p k as es a -> Doc x

namedTypeExports :: TSNamed 'Z k as es a -> Doc x

namedTypeExports_ :: TSNamed_ 'Z as es a -> Doc x

Serializing

encodeType :: TSType 'Z k a -> a -> ByteString

encodeTypeStrict :: TSType 'Z k a -> a -> ByteString

typeToValue :: TSType 'Z k a -> a -> Value

Parsing

decodeType :: TSType 'Z k a -> ByteString -> Either (ParseError ParseErr) a

decodeTypeStrict :: TSType 'Z k a -> ByteString -> Either (ParseError ParseErr) a

parseType :: forall k a. TSType 'Z k a -> Parse a